Logout and restructure code
