+++ package.json
@@ -5,7 +5,8 @@
   "dependencies": {
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
-    "react-google-login": "^4.0.0",
+    "react-facebook-login": "^4.1.1",
+    "react-google-login": "^4.0.1",
     "react-linkedin-sdk": "^1.0.10",
     "react-script-tag": "^1.1.2",
     "react-scripts": "1.1.4",
