+++ src/components/InfoScreen.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import "../style.css";
+
+class InfoScreen extends React.Component {
+  render() {
+    return (
+      <div className="contrastset">
+        <br />
+        <br />
+        <div className="center contact">Contact us!</div>
+        <br />
+        <div className="center siteref">
+          <a href="https://xane.ai">|| Xane.ai ||</a>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default InfoScreen;
