+++ src/config.json
@@ -1,5 +1,5 @@
 {
   "GOOGLE_CLIENT_ID": "935473844066-53q4l5k84ld9e2qo6bpjugttctrmhool.apps.googleusercontent.com",
-  "FACEBOOK_APP_ID": "",
+  "FACEBOOK_APP_ID": "338520096970139",
   "LINKEDIN_CLIENT_ID": "77xlojcghub00g"
 }
